added new controller for organization and better practices
